Update module Install and Remove Font

+++ FontManager/Manager/FileManager.cs
@@ -217,7 +217,7 @@ namespace FontManager
 
         public void DeleteFile(String filePath)
         {
-            if (!String.IsNullOrEmpty(filePath))
+            if (!String.IsNullOrEmpty(filePath) && File.Exists(filePath))
             {
                 File.Delete(filePath);
             }
